Can transmit and receive.
The only issue is the first receive after a transmit is missing the
first bit (so we get parity errors).

E.g. I expect 0xEE in response to an KEE command sent to a Keyboard
(it's the echo command), but I get 0xDC with a parity error. Need to
scope it out to see what's what.

Also removed the tests, because the big complicated PS2 class is gone
and there's much less to test now.